fix du main + diviser le fichier wildcard.c en deux fichier string_equal_wildcard et expand_wildcard
