+++ srcs/expansion/get_variable_key.c
