+++ srcs/expansion/string_equal_wildcard.c
@@ -1,27 +1,16 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   wildcard.c                                         :+:      :+:    :+:   */
+/*   string_equal_wildcard.c                            :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: brappo <brappo@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/04/11 12:39:55 by brappo            #+#    #+#             */
-/*   Updated: 2024/04/12 10:23:22 by brappo           ###   ########.fr       */
+/*   Updated: 2024/04/12 10:28:57 by brappo           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "minishell.h"
-
-static bool	add_word_token(t_list **head, char *str)
-{
-	t_token	*new_token;
-
-	new_token = t_token_init(ft_strdup(str), WORD);
-	if (lst_push_front_content(head,
-			new_token, t_token_free))
-		return (true);
-	return (false);
-}
 
 static size_t	get_next_wildcard(char *str_wildcard)
 {
@@ -41,7 +30,7 @@ static ssize_t	search_characters(char *characters,
 	char_pos = 0;
 	if (*characters == '*')
 	{
-		while (ft_strncmp(characters + 1, str + char_pos,
+		while (ft_strncmp(characters + 1, str + char_pos, \
 				character_length - 1))
 		{
 			if (str[char_pos] == '\0')
@@ -80,32 +69,4 @@ bool	string_equal_wildcard(char *str_wildcard, char *str_b)
 	if (*str_wildcard != '*' && *str_b)
 		return (false);
 	return (true);
-}
-
-t_list	*expand_wildcard(char *str)
-{
-	DIR				*current_directory;
-	struct dirent	*file;
-	t_list			*result;
-
-	current_directory = opendir(".");
-	if (current_directory == NULL)
-		return (NULL);
-	result = NULL;
-	file = readdir(current_directory);
-	if (file == NULL)
-		return (NULL);
-	while (file != NULL)
-	{
-		if (file->d_name[0] != '.' && string_equal_wildcard(str, file->d_name))
-		{
-			if (!add_word_token(&result, file->d_name))
-				return (ft_lstclear(&result, t_token_free), NULL);
-		}
-		file = readdir(current_directory);
-	}
-	closedir(current_directory);
-	if (result == NULL)
-		add_word_token(&result, str);
-	return (result);
 }
